update screenversion api

+++ app/src/main/java/com/example/myapplication/network/api/MainApi.java
@@ -43,7 +43,7 @@ public interface MainApi {
   @POST("/cms/api/v1/sdc/screen/code/generate")
   Call<GeneratePair> getGeneratePairCode();
   @GET("svms/api/v1/svc/screenversion/{id}")
-  Call<Integer> getScreenVersion(@Path("id") String id);
+  Call<Long> getScreenVersion(@Path("id") String id);
   @GET("cms/api/v1/sdc/screen/{screenId}/active-schedule")
   Call<ScreenScheduleResponse> getActiveSchedule(@Path("screenId") String screenId);
